Played around with widgets, decided to create 2 widgets that have the same functionality
reason to create two different classes is that each widget class has a defined classname
this makes it easier to style. I wanted to create one widget where you can define which classname
the widget gets.. But that wasn't possible or preferred.

But most of all.. it was an experience to create the widgets in the first place